Refactor UI components: rename 'ui' directory to 'gui' and update references
