+++ src/gui/highlighter.cpp
