+++ src/gui/highlighter.h
