+++ src/gui/mainwindow.h
